requirements.txt
Flask==1.0.2
Flask-Cache==0.13.1
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-speech==0.35.0
google-cloud-texttospeech==0.2.0
google-cloud-translate==1.3.1
googleapis-common-protos==1.5.3
Jinja2==2.10
pydub==0.22.1
requests==2.19.1
Werkzeug==0.14.1
